Renaming the service to config
